# src/main/java/com/letscome/cardgame/domain/dtos/GameDto.java
package com.letscome.cardgame.domain.dtos;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.persistence.*;

@Builder
@AllArgsConstructor
@NoArgsConstructor
@Data
public class GameDto {
    @JsonProperty(value = "user_id")
    private Long userId;
}