# src/main/java/com/letscome/cardgame/domain/dtos/UserResponseDetail.java
package com.letscome.cardgame.domain.dtos;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@AllArgsConstructor
@NoArgsConstructor
@Data
public class UserResponseDetail {
    @JsonProperty("user_id")
    private Long userId;
}